Fix: draw the license text centered on multi-screen setup
The current implementation of the license command uses the `s->top`
and `s->left` values, combined with the screen's height and width to
compute the center of the screen to draw on. This is erroneous since
the x and y values of the text being drawn are already relative to the
current screen, causing the text to be rendered off-screen.

Simply removing the top and left values from the computation fixes
this issue and allows the license text to be drawn centered on the
current screen, even in multi-screen setups with differently sized
screens.

Signed-off-by: Antoine Busque <antoinebusque@gmail.com>